requirements.txt
Django == 4.0.4
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
whitenoise==6.1.0 
Pillow==9.0.1
django-ckeditor==6.4.2
djangorestframework==3.13.1
django-environ==0.9.0
crispy-bootstrap5==0.6
django-crispy-forms==1.14.0
django-environ==0.9.0
